Add AbstractRange and AbstractRangeProduct model
